Rename Social Vulnerability to Social Vulnerability Score (#572)
* Rename Social Vulnerability to Social Vulnerability Score

* readd old files

* update requirements

* updated copyright

* fixed wrong caption

* renamded output name

* renamded output name in the code

* fixed typo in result name

* revert back to previous result name

* Changed copyright year

* modified copyright year
