+++ environment.yml
@@ -20,3 +20,4 @@ dependencies:
   - rtree>=1.1.0
   - scipy>=1.11.3
   - shapely>=2.0.2
+  - deprecated>=1.2.14
