+++ requirements.imports
@@ -15,3 +15,4 @@ shapely
 pycodestyle
 pytest
 python-jose
+Deprecated
