+++ requirements.txt
@@ -15,3 +15,4 @@ requests>=2.31.0
 rtree>=1.1.0
 scipy>=1.11.3
 shapely>=2.0.2
+Deprecated>=1.2.14
