+++ tests/pyincore/analyses/socialvulnerabilityscore/test_socialvulnerabilityscore.py
@@ -1,0 +1,34 @@
+# This program and the accompanying materials are made available under the
+# terms of the Mozilla Public License v2.0 which accompanies this distribution,
+# and is available at https://www.mozilla.org/en-US/MPL/2.0/
+
+
+from pyincore import IncoreClient
+from pyincore.analyses.socialvulnerabilityscore import SocialVulnerabilityScore
+import pyincore.globals as pyglobals
+
+
+def run_with_base_class():
+    census_geo_level = "tract"
+
+    # For a full example, datasets should be obtained using the Census API in pyincore_data
+    national_vulnerability_feature_averages = "6241d9c653302c512d67ef26"
+    # social_vulnerability_demographic_factors = "6241e58653302c512d67fb38"
+    social_vulnerability_demographic_factors = "62b4d2f7be53de4a4737e52d"  # calculated from censusutil
+
+    client = IncoreClient(pyglobals.INCORE_API_DEV_URL)
+    social_vulnerability_score = SocialVulnerabilityScore(client)
+
+    social_vulnerability_score.set_parameter("result_name", "social_vulnerabilty_score")
+
+    social_vulnerability_score.load_remote_input_dataset("national_vulnerability_feature_averages",
+                                                   national_vulnerability_feature_averages)
+    social_vulnerability_score.load_remote_input_dataset("social_vulnerability_demographic_factors",
+                                                   social_vulnerability_demographic_factors)
+
+    # Run pipeline damage analysis
+    result = social_vulnerability_score.run_analysis()
+
+
+if __name__ == "__main__":
+    run_with_base_class()
